fix some permissions in routes, add profile route abstract post to component.

+++ src/routes/posts/+page.server.ts
@@ -1,18 +1,25 @@
-import { createPost, getPosts, deletePost } from '$db/controllers/PostController';
+import { createPost, getPosts, deletePost, getPostById } from '$db/controllers/PostController';
 import type { RequestEvent, Action } from './$types.js';
 
-export const load = async (ctx: any) => {
+export const load = async ({ locals }: RequestEvent) => {
 	let posts = await getPosts();
 	posts = JSON.parse(JSON.stringify(posts));
-	// console.log(posts);
 	return {
-		posts
+		posts,
+		user: locals.authedUser
 	};
 };
 
 export const actions = {
 	create: async ({ request, locals }: RequestEvent) => {
 		const authedUser = locals.authedUser;
+
+		if (!authedUser) {
+			return {
+				status: 401
+			};
+		}
+
 		const data = await request.formData();
 		const title = data.get('title')?.toString() || '';
 		const content = data.get('content')?.toString() || '';
@@ -22,9 +29,20 @@ export const actions = {
 			status: res
 		};
 	},
-	delete: async ({ request }: RequestEvent) => {
+	delete: async ({ request, locals }: RequestEvent) => {
+		const authedUser = locals.authedUser;
+
 		const data = await request.formData();
 		const id = data.get('id')?.toString() || '';
+		let post = await getPostById(id);
+		post = JSON.parse(JSON.stringify(post));
+
+		if (!authedUser || !post || post.author._id !== authedUser._id) {
+			return {
+				status: 401
+			};
+		}
+
 		let res = await deletePost(id);
 		return {
 			status: res
